Move transcript code to transcript.rs
Closes #13
